implement responsiveness for footer card component
